+++ SM_Il2Cpp/Main.cs
@@ -25,7 +25,7 @@ namespace MelonLoader.Support
         private static ISupportModule_To Initialize(ISupportModule_From interface_from)
         {
             Interface = interface_from;
-            
+            MelonUtils.SetConsoleTitle(MelonUtils.GetVersionStrWithGameName(Application.version)); 
             UnityMappers.RegisterMappers();
 
             LogSupport.RemoveAllHandlers();